+++ SinZsEventTester/Framework/GSQTester.cs
@@ -9,6 +9,21 @@ namespace SinZsEventTester.Framework;
 /// <param name="reflector">SMAPI's reflection helper.</param>
 internal sealed class GSQTester(IMonitor monitor, IReflectionHelper reflector)
 {
+    private static Dictionary<string, Func<string, bool>> _additionalAssets = new(StringComparer.OrdinalIgnoreCase)
+    {
+        ["Data/MineCarts"] = static name => name is "MinecartsUnlocked",
+        ["Data/Characters"] = static name => name is "CanSocialize" or "CanVisitIsland" or "ItemDeliveryQuest" or "WinterStarParticipant" or "MinecartsUnlocked",
+};
+
+    internal static bool Register(IAssetName asset, Func<string, bool>? filter)
+        => _additionalAssets.TryAdd(asset.BaseName, filter ?? Extensions.IsPossibleGSQString);
+
+    internal static bool Register(IAssetName asset, HashSet<string> additional)
+        => _additionalAssets.TryAdd(asset.BaseName, name => Extensions.IsPossibleGSQString(name) || additional.Contains(name));
+
+    internal static bool Remove(IAssetName assets)
+        => _additionalAssets.Remove(assets.BaseName);
+
     private readonly SObject puffer = new("420", 1);
 
     /// <summary>
@@ -33,14 +48,40 @@ internal sealed class GSQTester(IMonitor monitor, IReflectionHelper reflector)
                     continue;
                 }
 
-                string[] breadcrumbs = [$"Data/{method.Name.Replace('_', '/')}"];
+                string asset = $"Data/{method.Name.Replace('_', '/')}";
+                string[] breadcrumbs = [asset];
 
-                this.Process(data, breadcrumbs);
+                this.Process(data, breadcrumbs, _additionalAssets.GetValueOrDefault(asset) ?? Extensions.IsPossibleGSQString);
             }
+        }
+
+        // create a new asset manager to avoid poisoning the one we're given.
+        LocalizedContentManager tempAssetManager = content.CreateTemporary();
+        try
+        {
+            foreach ((string asset, Func<string, bool> gsqfilter) in _additionalAssets)
+            {
+                if (asset.StartsWith("Data", StringComparison.OrdinalIgnoreCase))
+                {
+                    continue;
+                }
+
+                object? data = tempAssetManager.Load<object>(asset);
+                if (data is null)
+                {
+                    continue;
+                }
+
+                this.Process(data, [asset], gsqfilter);
+            }
+        }
+        finally
+        {
+            tempAssetManager.Dispose();
         }
     }
 
-    private void Process(object data, string[] breadcrumbs)
+    private void Process(object data, string[] breadcrumbs, Func<string, bool> filter)
     {
         if (data is null)
         {
@@ -64,12 +105,12 @@ internal sealed class GSQTester(IMonitor monitor, IReflectionHelper reflector)
             {
                 Type tkey = types.First();
                 MethodInfo processor = this.GetType().GetMethod(nameof(this.ProcessDictionary), BindingFlags.Instance | BindingFlags.NonPublic)!.MakeGenericMethod(tkey, dataType)!;
-                processor.Invoke(this, [data, breadcrumbs]);
+                processor.Invoke(this, [data, breadcrumbs, filter]);
             }
             else if (t.GetGenericTypeDefinition() == typeof(List<>))
             {
                 MethodInfo processor = this.GetType().GetMethod(nameof(this.ProcessList), BindingFlags.Instance | BindingFlags.NonPublic)!.MakeGenericMethod(dataType);
-                processor.Invoke(this, [data, breadcrumbs]);
+                processor.Invoke(this, [data, breadcrumbs, filter]);
             }
             else
             {
@@ -80,7 +121,7 @@ internal sealed class GSQTester(IMonitor monitor, IReflectionHelper reflector)
         {
             foreach (FieldInfo field in t.GetFields())
             {
-                if (field.Name.IsPossibleGSQString() && field.FieldType == typeof(string))
+                if (filter(field.Name) && field.FieldType == typeof(string))
                 {
                     string? gsq = (string?)field.GetValue(data);
                     this.CheckGSQ(gsq, [..breadcrumbs, field.Name]);
@@ -88,13 +129,13 @@ internal sealed class GSQTester(IMonitor monitor, IReflectionHelper reflector)
                 else if (!field.FieldType.IsValueType && (field.FieldType.IsGenericType || field.FieldType.Assembly!.GetName()!.Name!.Contains("StardewValley", StringComparison.OrdinalIgnoreCase)))
                 {
                     object f = field.GetValue(data)!;
-                    this.Process(f, [.. breadcrumbs, field.Name]);
+                    this.Process(f, [.. breadcrumbs, field.Name], filter);
                 }
             }
 
             foreach (PropertyInfo prop in t.GetProperties())
             {
-                if (prop.Name.IsPossibleGSQString() && prop.PropertyType == typeof(string))
+                if (filter(prop.Name) && prop.PropertyType == typeof(string))
                 {
                     string? gsq = (string?)prop.GetValue(data);
                     this.CheckGSQ(gsq, [..breadcrumbs, prop.Name]);
@@ -102,22 +143,22 @@ internal sealed class GSQTester(IMonitor monitor, IReflectionHelper reflector)
                 else if (!prop.PropertyType.IsValueType && (prop.PropertyType.IsGenericType || prop.PropertyType.Assembly!.GetName()!.Name!.Contains("StardewValley", StringComparison.OrdinalIgnoreCase)))
                 {
                     object f = prop.GetValue(data)!;
-                    this.Process(f, [.. breadcrumbs, prop.Name]);
+                    this.Process(f, [.. breadcrumbs, prop.Name], filter);
                 }
             }
         }
     }
 
-    private void ProcessDictionary<TKey, TValue>(Dictionary<TKey, TValue> data, string[] breadcrumbs)
+    private void ProcessDictionary<TKey, TValue>(Dictionary<TKey, TValue> data, string[] breadcrumbs, Func<string, bool> filter)
         where TKey : notnull
     {
         foreach ((TKey k, TValue v) in data)
         {
-            this.Process(v!, [.. breadcrumbs, k.ToString()!]);
+            this.Process(v!, [.. breadcrumbs, k.ToString()!], filter);
         }
     }
 
-    private void ProcessList<T>(List<T> data, string[] breadcrumbs)
+    private void ProcessList<T>(List<T> data, string[] breadcrumbs, Func<string, bool> filter)
     {
         for (int i = 0; i < data.Count; i++)
         {
@@ -131,7 +172,7 @@ internal sealed class GSQTester(IMonitor monitor, IReflectionHelper reflector)
             id ??= (reflector.GetProperty<string>(v, "Id", false) ?? reflector.GetProperty<string>(v, "ID", false))?.GetValue();
             id ??= i.ToString()!;
 
-            this.Process(v!, [.. breadcrumbs, id]);
+            this.Process(v!, [.. breadcrumbs, id], filter);
         }
     }
 
@@ -143,21 +184,58 @@ internal sealed class GSQTester(IMonitor monitor, IReflectionHelper reflector)
         }
 
         monitor.Log($"Checking: {gsq}\n{breadcrumbs.Render()}", LogLevel.Info);
-        GameStateQuery.CheckConditions(gsq, null, null, this.puffer, this.puffer, Random.Shared);
+
+        if (gsq is "TRUE" or "FALSE")
+        {
+            return;
+        }
+
+        Farmer? player = Game1.player;
+        GameLocation location = player?.currentLocation ?? Game1.currentLocation;
+        GameStateQuery.ParsedGameStateQuery[] parsed = GameStateQuery.Parse(gsq);
+        if (parsed.Length == 0)
+        {
+            return;
+        }
+
+        if (parsed[0].Error is { } error)
+        {
+            GameStateQuery.Helpers.ErrorResult(parsed[0].Query, error);
+            return;
+        }
+
+        foreach (GameStateQuery.ParsedGameStateQuery query in parsed)
+        {
+            // the ANY query is checked separately.
+            if (query.Query[0].Equals("ANY", StringComparison.OrdinalIgnoreCase))
+            {
+                foreach (string? subquery in query.Query.AsSpan(1))
+                {
+                    this.CheckGSQ(subquery, [.. breadcrumbs, gsq]);
+                }
+                continue;
+            }
+
+            try
+            {
+                query.Resolver(query.Query, location, player, this.puffer, this.puffer, Random.Shared);
+            }
+            catch (Exception ex)
+            {
+                monitor.Log($"Encountered exception running {string.Join(", ", query.Query)}, see log for details.", LogLevel.Error);
+                monitor.Log(ex.ToString());
+            }
+        }
     }
 }
 
+/// <summary>
+/// The extension methods for this class.
+/// </summary>
 file static class Extensions
 {
     internal static string Render(this string[] breadcrumbs) => string.Join("->", breadcrumbs);
 
     internal static bool IsPossibleGSQString(this string name)
-    {
-        if (name.EndsWith("Condition"))
-        {
-            return true;
-        }
-
-        return name is "CanSocialize" or "CanVisitIsland" or "ItemDeliveryQuest" or "WinterStarParticipant" or "MinecartsUnlocked";
-    }
+        => name.EndsWith("Condition");
 }