+++ SinZsEventTester/docs/Changelog.md
@@ -1,5 +1,9 @@
 ﻿Changelog
 ===============
 
-### Version 0.1.0
+### Version 0.1.2
+* Avoid shortcircuiting on GSQ checker.
+* Add way for mods to register their own assets that have GSQ.
+
+### Version 0.1.1
 * Initial upload